layers: Cleanup shader_instruction.cpp
